pyproject.toml
[project]
name = "ComfyUI-6yuan"
description = "A curated set of ComfyUI nodes under 6yuan namespace."
version = "1.0.0"
license = { text = "MIT" }
dependencies = [
    "transparent-background",
    "tqdm",
    "opencv-python",
    "scikit-learn",
    "numpy",
    "Pillow",
    "torch"
]

[project.urls]
Repository = "https://github.com/LvyuanW/ComfyUI-6yuan"

[tool.comfy]
PublisherId = "LvyuanW"
DisplayName = "ComfyUI-6yuan"
Icon = "💰"
